requirements.txt
flask==2.0.1
pyspark==3.4.0
pymongo==4.3.3
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
plotly==5.13.1
python-dotenv==1.0.0
flask-cors==4.0.0
pyarrow==11.0.0
kaleido==0.2.1
nbformat>=4.2.0 
pytest
openpyxl==3.1.2 